+++ nordea-eramaksu-configuration.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Define plugin default texts
+*/
+define("DESCRIPTION", "Nordea Finance Erämaksu on vakuudeton määräaikainen kertaluotto, joka on tarkoitettu isompiin, 500 – 30 000 euron suuruisiin verkko-ostoksiin.");
+define("LEGAL_INFORMATION", "Henkilötietojasi luovutetaan luotonantaja Nordea Rahoitus Suomi Oy:lle, joka käsittelee niitä rekisterinpitäjänä luoton myöntämistä ja valvontaa varten. Yksityiskohtaista tietoa henkilötietojen käsittelystä annetaan Nordean tietosuojaselosteessa, jonka voit saada myös ottamalla yhteyttä luotonantajaan. <a href='https://www.nordea.com/fi/tietosuojaseloste' target='_blank'>Tietosuojaseloste</a> sisältää tietoa henkilötietojesi käsittelyyn liittyvistä oikeuksistasi, joita ovat muun muassa oikeus tutustua tietoihin, oikeus tietojen oikaisemiseen ja oikeus siirtää tiedot järjestelmästä toiseen.");
+define("CUSTOM_TEXT", "Nordea Finance Erämaksu on vakuudeton määräaikainen kertaluotto, joka on tarkoitettu isompiin, 500 – 30 000 euron suuruisiin verkko-ostoksiin.");
+define("CONTINUE_SHOPPING_TEXT", "Jatka ostoksia");
+/**
+ * Define plugin API domain
+*/
+define("NORDEA_FINANCE_API_TEST_DOMAIN", "api.nordeaopenbanking.com");
+define("NORDEA_FINANCE_API_PRODUCTION_DOMAIN", "purchasefinance.nordeaopenbanking.com");
+?>